messages_service/requirements.txt
fastapi
uvicorn[standard]
hazelcast-python-client
pydantic
pydantic-factories